requirements.txt
django-configurations>=0.4
Django>=1.7,<1.8
dj-database-url==0.3.0

django-debug-toolbar>=1.2.1,<2.0
django_compressor==1.4
python-memcached==1.53

django-extensions==1.4.0
djangorestframework==2.4.2
https://github.com/plaid/plaid-python/archive/master.zip
twilio
pymongo

http://github.com/macropin/django-registration/archive/master.zip
celery[redis]
redis
numpy
scikit-learn
scipy
